// types.ts
export interface Address {
    province: string;
    district: string;
    subdistrict: string;
    postalCode: string;
    moo: string;
    street: string;
    soi: string;
    fullAddressText?: string; // Optional full text representation
}

// Interface for Promoter Information
export interface PromoterInfo {
    name: string;
    phone: string;
    email: string;
}

// Enum for Farmer Status
export enum FarmerStatus {
    APPROVED = 'อนุมัติ',
    PENDING = 'รอดำเนินการ',
    REJECTED = 'ไม่อนุมัติ',
}

// Interface สำหรับข้อมูลโปรไฟล์ผู้ใช้ (Updated)
export interface Profile {
  firstName: string;
  lastName: string;
  phone: string;
  email: string;
  lineId: string;
  address: Address;
  avatarUrl: string;
  status: FarmerStatus;
  promoterInfo?: PromoterInfo;
}